+++ Source/SelfService/Web/microservice/purchaseOrder/view.tsx
@@ -19,7 +19,7 @@ import { TabPanel } from '../../utils/materialUi';
 import EditIcon from '@mui/icons-material/Edit';
 
 import { DownloadButtons } from '../components/downloadButtons';
-import { HealthStatus } from '../microserviceView/healthStatus';
+import { HealthStatus } from '../microserviceView/healthStatus/healthStatus';
 
 type Props = {
     applicationId: string
