+++ Source/SelfService/Web/microservice/rawDataLog/view.tsx
@@ -7,7 +7,7 @@ import { useHistory } from 'react-router-dom';
 import { Stack } from '@fluentui/react/lib/Stack';
 import { Pivot, PivotItem } from '@fluentui/react';
 
-import { HealthStatus } from '../microserviceView/healthStatus';
+import { HealthStatus } from '../microserviceView/healthStatus/healthStatus';
 import { HttpResponsePodStatus } from '../../api/api';
 import { useReadable } from 'use-svelte-store';
 import { microservices, MicroserviceStore } from '../../stores/microservice';
